% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{.generate_enfa_summary_table}
\alias{.generate_enfa_summary_table}
\title{.generate_enfa_summary_table}
\usage{
.generate_enfa_summary_table(results, output_dir, verbose)
}
\arguments{
\item{results}{A list of ENFA results for the target and closest species.}

\item{output_dir}{Character. The directory to save the CSV.}

\item{verbose}{Logical. If TRUE, prints messages.}
}
\description{
Creates a CSV file comparing ENFA metrics for a target and its neighbors.
}
\keyword{internal}
